Module 5

+++ sick-fits/backend/src/resolvers/Mutation.js
@@ -1,13 +1,36 @@
 const bcrypt = require('bcryptjs');
 const jwt = require('jsonwebtoken');
+const { randomBytes } = require('crypto');
+const { promisify } = require('util');
+const { hasPermission } = require('../utils');
+const { transport, makeANiceEmail } = require('../mail');
+
+const signInUser = (ctx, userId) => {
+  const token = jwt.sign({ userId }, process.env.APP_SECRET);
+
+  ctx.response.cookie('token', token, {
+    httpOnly: true,
+    maxAge: 1000 * 60 * 60 * 24 * 365,
+  });
+};
 
 const Mutations = {
   async createItem(parent, args, ctx, info) {
-    // TODO: check if logged in
+    if (!ctx.request.userId) {
+      throw new Error('You must be logged inn to create item');
+    }
 
     const item = await ctx.db.mutation.createItem(
       {
-        data: { ...args },
+        data: {
+          user: {
+            // User to item relationship
+            connect: {
+              id: ctx.request.userId,
+            },
+          },
+          ...args,
+        },
       },
       info,
     );
@@ -33,7 +56,16 @@ const Mutations = {
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id };
 
-    const item = await ctx.db.query.item({ where }, `{ id title }`);
+    const item = await ctx.db.query.item({ where }, `{ id title user { id } }`);
+
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission =>
+      ['ADMIN', 'ITEMDELETE'].includes(permission),
+    );
+
+    if (!ownsItem && !hasPermissions) {
+      throw new Error('Not allowed to delete');
+    }
 
     return ctx.db.mutation.deleteItem({ where }, info);
   },
@@ -54,14 +86,119 @@ const Mutations = {
       info,
     );
 
-    const token = jwt.sign({ userId: user.id }, process.env.APP_SECRET);
-
-    ctx.response.cookie('token', token, {
-      httpOnly: true,
-      maxAge: 1000 * 60 * 60 * 24 * 365,
-    });
+    signInUser(ctx, user.id);
 
     return user;
+  },
+
+  async signin(parent, { email, password }, ctx, info) {
+    const user = await ctx.db.query.user({ where: { email } });
+
+    if (!user) {
+      throw new Error('Email or password wrong');
+    }
+
+    const valid = await bcrypt.compare(password, user.password);
+
+    if (!valid) {
+      throw new Error('Email or password wrong');
+    }
+
+    signInUser(ctx, user.id);
+
+    return user;
+  },
+
+  signout(parent, args, ctx, info) {
+    ctx.response.clearCookie('token');
+    return { message: 'Signout success' };
+  },
+
+  async requestReset(parent, { email }, ctx, info) {
+    const user = await ctx.db.query.user({ where: { email } });
+
+    if (!user) {
+      throw new Error(`The provided email (${email}) is not registered`);
+    }
+
+    const promisifiedRandomBytes = promisify(randomBytes);
+    const resetToken = (await promisifiedRandomBytes(20)).toString('hex');
+    const resetTokenExpiry = Date.now() + 3600000;
+
+    const res = await ctx.db.mutation.updateUser({
+      where: { email },
+      data: { resetToken, resetTokenExpiry },
+    });
+
+    const mailRes = await transport.sendMail({
+      from: '',
+      to: user.email,
+      subject: 'Password reset @ sick-fits',
+      html: makeANiceEmail(`
+        Click here to reset your password: \n\n
+        <a href="${
+          process.env.FRONTEND_URL
+        }/reset?resetToken=${resetToken}">Click here to Reset password</a>
+      `),
+    });
+
+    return { message: 'Sent reset mail' };
+  },
+
+  async resetPassword(parent, args, ctx, info) {
+    const { resetToken, password, confirmPassword } = args;
+
+    if (password !== confirmPassword) {
+      throw new Error('Passwords does not match');
+    }
+
+    const [user] = await ctx.db.query.users({
+      where: {
+        resetToken,
+        resetTokenExpiry_gte: Date.now() - 3600000,
+      },
+    });
+
+    if (!user) {
+      throw new Error('Invalid or expired resetToken');
+    }
+
+    const newPassword = await bcrypt.hash(password, 10);
+
+    const updatedUser = await ctx.db.mutation.updateUser({
+      where: { email: user.email },
+      data: {
+        password: newPassword,
+        resetToken: null,
+        resetTokenExpiry: null,
+      },
+    });
+
+    signInUser(ctx, updatedUser.id);
+
+    return updatedUser;
+  },
+
+  async updatePermissions(parent, args, ctx, info) {
+    if (!ctx.request.userId) {
+      throw new Error('You must be logged in to perform this action');
+    }
+
+    hasPermission(ctx.request.user, ['ADMIN', 'PERMISSIONUPDATE']);
+
+    return ctx.db.mutation.updateUser(
+      {
+        data: {
+          permissions: {
+            set: args.permissions,
+          },
+        },
+        where: {
+          id: args.userId,
+        },
+      },
+      info,
+    );
   },
 };
